refactor: remove mapstruct, impl static mapping methods

+++ subscription-service/src/main/java/com/dmytrocherkes/subscriptionservice/repository/SubscriptionRepository.java
@@ -9,5 +9,7 @@ import java.util.UUID;
 
 @Repository
 public interface SubscriptionRepository extends JpaRepository<Subscription, UUID> {
-    List<Subscription> findAllByUserId(Long userId);
+
+    List<Subscription> findAllByUserId(Integer userId);
+
 }
